Added checkbox options; female by default

+++ server.R
@@ -1,34 +1,40 @@
 library(shiny)
-library(digest)
 
-## Titles: 
-
-# Define server logic required to draw a histogram
 shinyServer(function(input, output) {
   
   gridData <- reactiveValues();
   gridData$Labels <- NULL;
   gridData$Dims <- NULL;
-
-  output$myImage <- renderImage({
-    if(is.null(input$picBase)){
-      return();
-    }
-    # A temp file to save the output.
-    # This file will be removed later by renderImage
-    outdir <- tempdir();
+  
+  makeFile <- function(outfile){
     infile <- tempfile(fileext='.tex');
-    outfile <- tempfile(fileext='.png');
+    outdir <- tempdir();
     infileBase <- sub("\\.tex$","",infile);
-    outfileBase <- sub("\\.png$","",outfile);
-    tikzString <- paste(input$picBase,"female","monitor","sword","shirt=red",sep=",");
-    personName <- ifelse(input$cName=="",
-                         "Audrey Hacker",input$cName);
-    personJob <- ifelse(input$dayJob=="",
-                        "Conflict Negotiator",input$dayJob);
-    personSkills <- gsub("\n","\\\\\n",input$skills);
+    outfileBase <- sub("\\.(png|pdf)$","",outfile);
+    faceOptions <- debounce(reactive({input$faceOptions}), 500);
+    cName <- debounce(reactive({input$cName}), 500);
+    dayJob <- debounce(reactive({input$dayJob}), 500);
+    skill <- debounce(reactive({input$skill}), 500);
+    optionString <- c(faceOptions(),"female");
+    if("male" %in% optionString){
+      optionString <- setdiff(optionString, c("male", "female"));
+    }
+    print(optionString);
+    tikzString <- sub(",$","",paste(input$picBase,optionString,sep=",",
+                                    collapse=","));
+    print(tikzString);
+    personName <- ifelse(cName()=="",
+                         "Audrey Hacker",cName());
+    personJob <- ifelse(dayJob()=="",
+                        "Conflict Negotiator",
+                        dayJob());
+    personSkill <- skill();
+    print(personSkill);
+    personName <- sub("[%\\].*$","",personName);
+    personJob <- sub("[%\\].*$","",personJob);
+    personSkill <- sub("[%\\].*$","",personSkill);
     cat(
-"     \\documentclass[11pt,a4paper]{memoir}
+      "     \\documentclass[11pt,a4paper]{memoir}
       
       \\setstocksize{55mm}{85mm} % UK Stock size
       \\setpagecc{55mm}{85mm}{*}
@@ -36,11 +42,11 @@ shinyServer(function(input, output) {
       \\setulmargins{5mm}{*}{*}
       \\setlrmargins{5mm}{*}{*}
       \\usepackage{xcolor}
-
+      
       \\nonstopmode
       \\setheadfoot{0.1pt}{0.1pt}
       \\setheaderspaces{1pt}{*}{*}
-
+      
       \\checkandfixthelayout[fixed]
       
       \\pagestyle{empty}
@@ -49,52 +55,65 @@ shinyServer(function(input, output) {
       \\usepackage{xcolor}
       \\usepackage{tikzpeople}
       \\usepackage{tabularx}
-
-\\newcolumntype{L}{>{\\arraybackslash}p{3.5cm}}
-
+      
+      \\newcolumntype{L}{>{\\arraybackslash}p{3.5cm}}
+      
       \\begin{document}
       \\noindent",
-sprintf("\\textbf{%s}\\\\",personName),
+      sprintf("\\textbf{%s}\\\\",personName),
       sprintf("\\tiny %s\\\\", 
               "GovHacker"),"
-      \\rule{74mm}{.3mm}\\\\
+      \\rule{65mm}{.3mm}\\\\
       \\begin{minipage}[t]{20mm}
       \\vspace{2mm}%",
       paste0("\\tikz{\\node[",tikzString,
-             ",minimum height=25mm]{}}"),
-"     \\end{minipage}
+             ",minimum height=20mm]{}}"),
+      "     \\end{minipage}
       \\hspace{1mm}
       \\begin{minipage}[t]{47mm}
       \\vspace{-0mm}%
       \\begin{flushleft}
       {
-       \\
-       \\begin{tabular}{lL}",
+      \\
+      \\begin{tabular}{lL}",
         sprintf("{\\color{gray}Day Job} & %s\\\\",personJob),
         "\\rule{0pt}{0.5em}\\\\",
-        sprintf("\\color{gray}Hacker Skills} & \\multicolumn{1}{m{3.5cm}}{%s}\\\\",personSkills),"
-        
-        \\vspace*{2mm}
-        \\end{tabular}
-      }
+        sprintf("\\color{gray}Hacker Skill} & \\multicolumn{1}{m{3.5cm}}{%s}\\\\",personSkill),"
+      
+      \\vspace*{2mm}
+      \\end{tabular}
+  }
       \\end{flushleft}
       \\end{minipage}
       \\end{Spacing}
       \\end{document}
-",
+      ",
       file=infile, sep="\n");
     system(command = paste("pdflatex","-output-directory",
                            outdir,infile,"-interaction batchmode"),
            ignore.stdout = TRUE);
-    system(command = paste("pdftoppm -png",
-                           "-singlefile", paste0(infileBase,".pdf"),
-                           "-scale-to-x",850,
-                           "-scale-to-y",550,outfileBase));
     unlink(paste0(infileBase,".tex"));
     unlink(paste0(infileBase,".aux"));
     unlink(paste0(infileBase,".log"));
-    unlink(paste0(infileBase,".pdf"));
-    gridData$fileName <- outfile;
+    if(grepl("\\.png$", outfile)){
+      system(command = paste("pdftoppm -png",
+                             "-singlefile", paste0(infileBase,".pdf"),
+                             "-scale-to-x",850,
+                             "-scale-to-y",550,outfileBase));
+      unlink(paste0(infileBase,".pdf"));
+    } else {
+      file.rename(paste0(infileBase,".pdf"), outfile);
+    }
+  }
+
+  output$myImage <- renderImage({
+    if(is.null(input$picBase)){
+      return();
+    }
+    # A temp file to save the output.
+    # This file will be removed later by renderImage
+    outfile <- tempfile(fileext='.png');
+    makeFile(outfile);
     
     # Return a list containing the filename
     list(src = outfile,
@@ -102,7 +121,7 @@ sprintf("\\textbf{%s}\\\\",personName),
          width = 425,
          height = 275,
          alt = "This is alternate text")
-  }, deleteFile = FALSE)
+  }, deleteFile = TRUE)
   
      
   output$cardText <- renderUI({
@@ -121,30 +140,29 @@ sprintf("\\textbf{%s}\\\\",personName),
   
   output$hackerCard.pdf <- downloadHandler(
     filename = function(){
-      fName <- sprintf("hackerCard_%s_%s.pdf",input$cName,
+      personName <- ifelse(input$cName=="",
+                           "Audrey Hacker",input$cName);
+      fName <- sprintf("hackerCard_%s_%s.pdf",personName,
                        format(Sys.Date(),"%Y-%b-%d"));
-      cat("Writing to file: ",fName, "\n");
       return(fName);
       },
     content = function(con){
-      pdf(con, width=8, height=8);
-      drawGrid();
-      dev.off();
+      makeFile(con);
     },
     contentType = "text/pdf"
   );
   
   output$hackerCard.png <- downloadHandler(
     filename = function(){
-      fName <- sprintf("hackerCard_%s_%s.png",input$cName,
+      personName <- ifelse(input$cName=="",
+                           "Audrey Hacker",input$cName);
+      fName <- sprintf("hackerCard_%s_%s.png",personName,
                        format(Sys.Date(),"%Y-%b-%d"));
       cat("Writing to file: ",fName, "\n");
       return(fName);
     },
     content = function(con){
-      png(con, width=1024, height=1024, pointsize = 18);
-      drawGrid();
-      dev.off();
+      makeFile(con);
     },
     contentType = "image/png"
   );
